move more files to ts
