+++ src/components/App.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import TopHeader from './TopHeader';
+import TopHeader from './TopHeader.tsx';
 
 const App = ({ children }) => (
   <div>
