+++ typings/redux-form-material-ui.d.ts
@@ -1,0 +1,9 @@
+declare module "redux-form-material-ui" {
+    import { ComponentType } from "react";
+
+    export interface TextFieldProps {
+        floatingLabelText: String;
+    }
+
+    export const TextField: ComponentType<TextFieldProps>;
+}
